requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
pydantic==2.5.3
jinja2==3.1.3
openai>=1.50.0
httpx>=0.27.0,<0.28.0
python-dotenv==1.0.0
pdfplumber==0.11.0
python-multipart>=0.0.6
